+++ QuickSchedule/WorkShift.m
@@ -48,7 +48,7 @@
     NSInteger ti = (NSInteger)interval;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hour = (ti / 3600);
-    return [NSString stringWithFormat:@"%i:%02i", hour, minutes];
+    return [NSString stringWithFormat:@"%li:%02li", (long)hour, (long)minutes];
 }
 
 @end
